requirements.txt
backcall
certifi
chardet
decorator
Django
gunicorn
idna
ipython
ipython-genutils
jedi
parso
pexpect
pickleshare
prompt-toolkit
psycopg2
ptyprocess
Pygments
pytz
requests
six
sqlparse
traitlets
urllib3
wcwidth
djangorestframework-simplejwt
django_rest
djangorestframework
djoser
drf_spectacular
django-extensions